remove some IE compat hacks that seem to break more than they fix. This makes our behaviour more similar to the way moz behaves, and fixes beneath others the about page.
svn path=/trunk/kdelibs/; revision=177992

+++ khtml/rendering/render_table.h
@@ -137,7 +137,7 @@ public:
     int colToEffCol( int col ) const {
 	int c = 0;
 	int i = 0;
-	while ( c < col && i < columns.size() ) {
+	while ( c < col && i < (int)columns.size() ) {
 	    c += columns[i].span;
 	    i++;
 	}
